fixes: sorting didn't evoke update, demo folder is missing by first launching

+++ de_dittner_mp3book/playlist/domain/service/ValidationDomainService.swift
@@ -32,7 +32,7 @@ class ValidationDomainService {
                 if !url.fileExists() {
                     return false
                 }
-            } else if let id = UInt64(b.playlistID), !LibraryContext.shared.iPodAppService.playlistExists(persistentID: id) {
+            } else if let id = b.playlistID, !LibraryContext.shared.iPodAppService.playlistExists(persistentID: id) {
                 return false
             }
         }
